- Fix effect attribute look up errors - Preliminary support for GBK encoding in XML parser (working in progress)

+++ garnet3d/main/src/priv/gfx/gpu/d3d11/d3d11Shader.h
@@ -76,8 +76,8 @@ namespace GN { namespace gfx
 
     struct D3D11AttributeParameterDesc : public GpuProgramAttributeParameterDesc
     {
-        StrA   semanticName;
-        size_t semanticIndex;
+        StrA semanticName;
+        UINT semanticIndex;
 
         /// ctor
         D3D11AttributeParameterDesc()
